put hashing functions to util modul
